PydanticAI_Agents/BasicUsage/requirements.txt
pydantic-ai==0.0.15
openai==1.58.1
colorama==0.4.6
dataclasses==0.6
asyncio==3.4.3
devtools==0.12.2
python-dotenv
nest-asyncio